make convert working with Number as well
