+++ src/pyuncertainnumber/characterisation/uncertainNumber.py
@@ -20,6 +20,8 @@ from numbers import Number
 from ..pba.distributions import Distribution as pbaDistribution
 import operator
 from pint import Quantity
+from ..gutils import is_un, pass_down_units
+
 
 """ Uncertain Number class """
 
@@ -734,54 +736,6 @@ class ParamSpecification:
             self._true_type = "distribution"
         else:
             self._true_type = "pbox"
-
-
-# * ---------------------helper functions  --------------------- *#
-
-
-def pass_down_units(a, b, ops, t):
-    """pass down the unit of the uncertain number
-
-    args:
-        - a: the first uncertain number
-        - b: the second uncertain number
-        - ops: the operation to be performed
-        - t: the result uncertain number of the operation
-    """
-    if is_un(b) == 0:
-        try:
-            new_q = ops(a._physical_quantity, b * a._physical_quantity.units)
-        except Exception:
-            new_q = ops(a._physical_quantity, b)
-    elif is_un(b) == 1:
-        new_q = ops(a._physical_quantity, b._physical_quantity)
-
-    t.physical_quantity = new_q
-    return t
-
-
-def is_un(sth):
-    """utility function to decide the essence of the object
-
-    returns:
-        - 0: if sth is a regular number, float or int
-        - 1: if sth is an UncertainNumber object
-        - 2: if sth is a construct in {Interval, Pbox, DempsterShafer, or Distribution}
-    """
-    from pyuncertainnumber import Interval, Pbox, DempsterShafer
-    from ..pba.distributions import Distribution
-
-    if isinstance(sth, Number):
-        return 0
-    elif isinstance(sth, UncertainNumber):
-        return 1
-    elif isinstance(sth, Interval | Pbox | DempsterShafer | Distribution):
-        return 2
-
-
-def exist_un(a_list) -> bool:
-    """check if there is any UN object in the list"""
-    return any(is_un(x) == 1 for x in a_list)
 
 
 # * ---------------------parametric shortcuts  --------------------- *#
